improve prep and clean up

+++ test/commands/hello/org.test.ts
@@ -1,8 +1,9 @@
 import * as fs from 'fs';
-import { exec } from 'child_process';
+import { exec, ExecException } from 'child_process';
 import * as assert from 'assert';
+import * as rimraf from 'rimraf';
 
-async function myExec(cmd:string) {
+async function myExec(cmd:string):Promise<{err: ExecException | null, stdout: string, stderr: string}> {
   return new Promise((resolve, reject) => {
     setTimeout(() => reject(), 10000);
     exec(cmd, (err, stdout, stderr) => {
@@ -11,28 +12,43 @@ async function myExec(cmd:string) {
   });
 }
 
-function turnGitOn(dir: string) {
-  process.chdir(dir);
-  fs.renameSync('.notgit','.git');
+function prep(projectPath: string) {
+  if (!fs.existsSync(projectPath)) {
+    console.error(`${projectPath} project path not found for prep, tests results not valid`)
+  } else {
+    process.chdir(projectPath);
+    if (fs.existsSync('.notgit')) {
+      fs.renameSync('.notgit','.git');
+    }
+    if (fs.existsSync('deploy')) {
+      rimraf('deploy', (err) => {});
+    }
+  }
 }
 
-function turnGitOff() {
-  fs.renameSync('.git','.notgit');
+function cleanUp(projectPath: string) {
+  if (fs.existsSync('.git')) {
+    fs.renameSync('.git','.notgit');
+  }
+  if (fs.existsSync('deploy')) {
+    rimraf('deploy', (err) => {});
+  }
 }
 const proj_path = "test/projects/basic_change";
 describe('git:package', () => {
   before(() => {
-    turnGitOn(proj_path);
+    prep(proj_path);
   });
   it('it builds a deployment with changed files', async () => {
     try {
       const res = await myExec('sfdx git:package -d deploy --purge');
+      assert.equal(null, res.err);
       assert.ok(fs.existsSync("deploy/package.xml"));
     } catch (e) {
       assert.fail(e);
     }
   });
   after(() => {
-    turnGitOff();
+    cleanUp(proj_path);
   });
 });
